+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to the FolioBlocks project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.10] - 2025-12-23
+### Fixed
+- Bug in Modular Gallery border controls.
+- Bug in playback controls on Carousel Gallery Block.
+
 ## [1.0.9] - 2025-12-18
 ### Fixed 
 - Bug in Masonry Gallery Image Filtering